requirements.txt
Django>=5.2.8
djangorestframework>=3.16.1
django-cors-headers>=4.9.0
channels>=4.0.0
channels-redis>=4.2.0

